requirements.txt
enum==0.4.4
Flask==0.10.1
Flask-Redis==0.0.6
gevent==1.0.1
greenlet==0.4.3
grizzled-python==1.0.1
itsdangerous==0.24
Jinja2==2.7.3
lxml==3.3.5
Markdown==2.5.2
MarkupSafe==0.23
misaka==1.0.2
prettytable==0.7.2
pycrypto==2.6.1
Pygments==2.0.2
PyYAML==3.11
redis==2.10.3
sae-python-dev==1.3.2
sqlcmd==0.7.1
Werkzeug==0.9.6
gunicorn==0.17.2
qiniu==7.0.2
